refactor: component 폴더 구조 세분화
